+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 5.0.1-dev.1
+
+* Add option for iOS/MacOS to allow non-biometric authentication (`darwinBiometricOnly`) #101
+  * Improve [canAuthenticate] to differentiate between no available biometry and no available 
+    user code.
+
 ## 5.0.0+4
 
 * Add topics to pubspec.yaml